Feature/file handling (#48)
* Added File service and Document file service implementations.

* Added file nadling and integrated it into existing codebase. Added almost all tests.

* Added tests for added methods in involvement service.

* Made filesystem save directory customizable from properties file. Added File Service tests.

* Fixed test failing bug regarding file data storage. Fixed javax validation bug for documents and involvements.

* Fixed security flaw enabling user to edit other people's involvements. Optimized ES repository scanning.

* Added indexing and content parsing with language recognition for pdf files.

* Enabled reindexing when editing existing documents.

* Fixed failing test.

* Added ICU tokenizer installation command in github workflow configuration.

* Added plugin installation into separate step.

* Fixed order of steps.

* Fixed syntax error.

* Set ES test container name explicitly.

* Update workflow.yml

Added debug print docker ps

* Update workflow.yml

added name option in es service

* Update workflow.yml

Added docker restart after exec icu install

* Added tests.

* Naming changes, code simplification and decoupling. Moved language models loading in bean configuration.

---------

+++ src/main/java/rs/teslaris/core/model/document/DocumentFile.java
@@ -35,7 +35,7 @@ public class DocumentFile extends BaseEntity {
     private String mimeType;
 
     @Column(name = "file_size", nullable = false)
-    private int fileSize;
+    private Long fileSize;
 
     @Column(name = "resource_type", nullable = false)
     private ResourceType resourceType;
